Move shared errors to shared/errors/ folder
