+++ src/core/index.js
@@ -1,0 +1,2 @@
+export * from './term-application'
+export * from './command'
