+++ src/main.js
@@ -1,12 +1,7 @@
-export * from './core/term-application'
-export * from './core/command'
-
+export * from './core'
 export * from './commands'
-
-export * from './services/file.service'
-export * from './services/exec/exec.service'
-export * from './services/deploy/deploy.service'
-export * from './services/config/config.service'
+export * from './services'
 
 export * from './shared/utils'
 export * from './shared/config'
+export * from './shared/errors'
