--- src/services/config/index.js
@@ -1,2 +1,0 @@
-export * from './config.service'
-export * from './config-file-not-exist.error'
