--- src/services/deploy/index.js
@@ -1,2 +1,0 @@
-export * from './deploy.service'
-export * from './target-file-already-exist.error'
