--- src/services/exec/index.js
@@ -1,3 +1,0 @@
-export * from './exec.service'
-export * from './bad-script-permission.error'
-export * from './script-not-exist.error'
