+++ src/services/index.js
@@ -1,4 +1,4 @@
-export * from './config'
-export * from './deploy'
-export * from './exec'
+export * from './config.service'
+export * from './deploy.service'
+export * from './exec.service'
 export * from './file.service'
