+++ src/shared/errors/bad-script-permission.error.js
