+++ src/shared/errors/script-not-exist.error.js
